feat: full scenescript output parsing

+++ src/room/bbox.rs
@@ -2,14 +2,14 @@ use crate::utils::vec3::Vec3;
 
 #[derive(Debug)]
 pub struct Bbox {
-    id: usize,
+    id: u64,
     position: Vec3,
     angle: f64,
     scale: Vec3,
 }
 
 impl Bbox {
-    pub fn new(id: usize, position: Vec3, angle: f64, scale: Vec3) -> Self {
+    pub fn new(id: u64, position: Vec3, angle: f64, scale: Vec3) -> Self {
         Self {
             id,
             position,
